pyproject.toml
[tool.poetry]
name = "dj-resturant-ms"
version = "0.1.0"
description = ""
authors = ["Mohsin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.3"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
twitter = "^1.19.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
colorlog = "^6.7.0"
requests = "^2.31.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
